+++ cpp/src/main.cpp
@@ -54,7 +54,7 @@ uint32_t random_id() {
     return id;
 }
 
-constexpr const char* VERSION = "1.4.1";
+constexpr const char* VERSION = "1.4.2";
 constexpr const char* COPYRIGHT = "Copyright (c) 2026 Mesh Console contributors. MIT License.";   // <- put your name here
 
 void usage() {